% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_vg_sales.R
\docType{data}
\name{Dados}
\alias{Dados}
\title{Dados de vendas sobre os jogos de video games}
\format{
Um objeto classificado como data frame, contendo 16719 observações e 10 variáveis.
}
\usage{
Dados
}
\arguments{
\item{Name:}{Nome do jogo}

\item{Platform:}{Plataforma do jogo}

\item{Year_of_Release:}{Ano de lançamento}

\item{Genre:}{Gênero do jogo}

\item{Publisher:}{Editora}

\item{NA_Sales:}{Vendas na América do Norte}

\item{EU_Sales:}{Vendas na União Europeia}

\item{JP_Sales:}{Vendas no Japão}

\item{Other_Sales:}{Vendas em outras localidades.}

\item{Global_Sales:}{Total de vendas entre os Países}

\item{Rating:}{Rank do jogo}
}
\value{
Retorna um banco de dados com os nome dos jogos,
as plataformas, o ano de lançamento, gênero do jogos
as editoras de jogos e as Vendas nas localidades.
}
\description{
Este conjunto de dados é sobre as vendas de jogos ao longo
dos anos em diversos países/continentes, e demais detalhes.
}
\examples{
# Vizualização geral dos dados
dplyr::glimpse(Dados)

}
\keyword{datasets}
