% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_series.R
\name{df_series}
\alias{df_series}
\title{Vendas de jogos por ano e coninentes.}
\usage{
df_series(db, Gen, Plat, Data_inicio, Data_final)
}
\arguments{
\item{db:}{Banco de dados, funcional para `Dados`.}

\item{Gen:}{Gênero, se prefirir todos, o argumento deve ser "Todos".}

\item{Plat:}{Plataforma, se prefirir todos, o argumento deve ser "Todos".}

\item{Data_inicio:}{Data inicial do intervalo da serie do banco de dados.}

\item{Data_final:}{Data final do intervalo da serie do banco de dados.}
}
\description{
Uma função que retorna um conjunto de dados da série sobre as vendas de jogos
acumalado para cada localidade por Ano. Pode-se restringir a Plataforma e o
Gênero de preferencia do úsuario.
}
\examples{
# Banco de dados `Dados` do pacote recomendado para utilizar nessa função.
# Segue abaixo, alguns valores dos demais parametros de entradas:
df_series(Dados, "Todos", "Todos", 2000, 2010)
}
