% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBoots_Vendas.R
\name{IBoots_Vendas}
\alias{IBoots_Vendas}
\title{Intervalo Bootstrap para vendas}
\usage{
IBoots_Vendas(n)
}
\arguments{
\item{n}{numero de reamostragem bootstrap.}
}
\value{
Um data.frame contendo um intervalo superior e inferior de 95% e a mediana para
cada região.
}
\description{
Função que recebe um quantidade de numero de reamostragem bootsrap e com as
seguintes seguintes regiões do banco de dados, demostra os seus resultados em
um data frame.
}
\examples{
IBoots_Vendas(100)

}
