% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_medias.R
\name{ic_medias}
\alias{ic_medias}
\title{Intervalo de confiança para vendas}
\usage{
ic_medias(n, amostra)
}
\arguments{
\item{x}{um vetor que contém número de vendas}

\item{alpha}{Nível de significancia do intervalo de confiança.}
}
\value{
Um vetor  contendo , intervalo superior e inferior do ic de vendas, e a media para cada categoria.
}
\description{
´
Função recebe um vetor com vendas por categorias e retorna o intervalo de confiança bootstrap, para a média
de vendas para cada uma das categorias.
}
