% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBoots_Vendas.R
\name{IBoots_Vendas}
\alias{IBoots_Vendas}
\title{Intervalo Bootstrap para vendas}
\usage{
IBoots_Vendas(n, amostra)
}
\arguments{
\item{n}{numero de reamostragem bootstrap.}

\item{amostra}{amostra de vendas de uma região, argumento deve ser um vetor.}
}
\value{
Um vetor contendo um intervalo superior e inferior de 95% e a mediana.
}
\description{
Função recebe um vetor com vendas por categorias e retorna o intervalo de bootstrap, para a média
de vendas para cada uma das regiões de vendas.
}
\examples{
# O Banco de dados precisa estar classificado como 'data.frame',
# isso é, esse é o motivo de usar a função "as.data.frame()"
IBoots_Vendas(100, as.data.frame(Dados)$NA_Sales)

}
